requirements.txt
asgiref==3.9.1
Django==5.2.4
django-decouple==2.1
django-tinymce==4.1.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
sqlparse==0.5.3
